Separate base wave model from the additive synth model.
